전체적 구조 변경, Test 로직 및 서순 개선
전체 패키지가 너무 혼잡한 듯 하여 `task`와 `catToy`패키지로 나눠 구성했습니다.

Test들의 내용과 로직, 설명을 개선하였습니다.
TaskControllerWebTest 마지막의 `deleteMvc` 부분 실행 시
`com.codesoom.assignment.task.TaskNotFoundException`
단 한 줄의 에러가 발생하는 부분이 있어 수정 중 입니다.

질문사항 : WebMvcTest에서 `TaskService`만 사용하는데 `CatToyService`까지 MockBean을 만들어야 실행이 되던데 이유의 대한 힌트를 주실 수 있으신가요
