--- app/src/main/java/com/codesoom/assignment/TaskNotFoundException.java
@@ -1,7 +1,0 @@
-package com.codesoom.assignment;
-
-public class TaskNotFoundException extends RuntimeException {
-    public TaskNotFoundException(Long id) {
-        super("Task not found: " + id);
-    }
-}
