--- app/src/main/java/com/codesoom/assignment/ToyNotFoundException.java
@@ -1,7 +1,0 @@
-package com.codesoom.assignment;
-
-public class ToyNotFoundException extends RuntimeException {
-    public ToyNotFoundException(Long id) {
-        super("Cat\'s Toy not found : ID = "+ id);
-    }
-}
