+++ app/src/main/java/com/codesoom/assignment/catToy/ToyNotFoundException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.catToy;
+
+public class ToyNotFoundException extends RuntimeException {
+    public ToyNotFoundException(Long id) {
+        super("Cat\'s Toy not found : ID = "+ id);
+    }
+}
