+++ app/src/main/java/com/codesoom/assignment/catToy/dto/ErrorResponse.java
@@ -1,0 +1,13 @@
+package com.codesoom.assignment.catToy.dto;
+
+public class ErrorResponse {
+    private String message;
+
+    public ErrorResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
