--- app/src/main/java/com/codesoom/assignment/domain/Task.java
@@ -1,30 +1,0 @@
-package com.codesoom.assignment.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-public class Task {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    private String title;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-}
