--- app/src/main/java/com/codesoom/assignment/domain/TaskRepository.java
@@ -1,15 +1,0 @@
-package com.codesoom.assignment.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface TaskRepository {
-
-    List<Task> findAll();
-
-    Optional<Task> findById(Long id);
-
-    Task save(Task task);
-
-    void delete(Task task);
-}
