+++ app/src/main/java/com/codesoom/assignment/task/TaskNotFoundException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.task;
+
+public class TaskNotFoundException extends RuntimeException {
+    public TaskNotFoundException(Long id) {
+        super("Task not found: " + id);
+    }
+}
