+++ app/src/main/java/com/codesoom/assignment/task/controllers/TaskController.java
@@ -1,0 +1,52 @@
+package com.codesoom.assignment.task.controllers;
+
+import com.codesoom.assignment.task.application.TaskService;
+import com.codesoom.assignment.task.domain.Task;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/tasks")
+//@CrossOrigin
+public class TaskController {
+
+    private TaskService taskService;
+
+    public TaskController(TaskService taskService) {
+        this.taskService = taskService;
+    }
+
+    @GetMapping
+    public List<Task> list() {
+        return taskService.getTasks();
+    }
+
+    @GetMapping("{id}")
+    public Task detail(@PathVariable Long id) {
+        return taskService.getTask(id);
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public Task create(@RequestBody Task task) {
+        return taskService.createTask(task);
+    }
+
+    @PutMapping("{id}")
+    public Task update(@PathVariable Long id, @RequestBody Task task) {
+        return taskService.updateTask(id, task);
+    }
+
+    @PatchMapping("{id}")
+    public Task patch(@PathVariable Long id, @RequestBody Task task) {
+        return taskService.updateTask(id, task);
+    }
+
+    @DeleteMapping("{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void delete(@PathVariable Long id) {
+        taskService.deleteTask(id);
+    }
+}
