+++ app/src/main/java/com/codesoom/assignment/task/controllers/TaskErrorAdvice.java
@@ -1,0 +1,19 @@
+package com.codesoom.assignment.task.controllers;
+
+import com.codesoom.assignment.task.TaskNotFoundException;
+import com.codesoom.assignment.task.dto.ErrorResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ControllerAdvice
+public class TaskErrorAdvice {
+    @ResponseBody
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(TaskNotFoundException.class)
+    public ErrorResponse handleNotFound() {
+        return new ErrorResponse("Task not found");
+    }
+}
