+++ web/package.json
@@ -35,7 +35,7 @@
     "given2": "^2.1.7",
     "jest": "^26.6.3",
     "jest-plugin-context": "^2.9.0",
-    "playwright": "^1.8.1",
+    "playwright": "^1.11.1",
     "redux-mock-store": "^1.5.4",
     "start-server-and-test": "^1.12.0",
     "webpack": "^5.21.2",
